requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.5.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
instructor==1.10.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
numpy==2.2.3
openai==1.97.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.1
requests==2.32.3
rich==14.0.0
scipy==1.16.1
selenium==4.34.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.7
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
torch==2.7.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xlrd==2.0.1
yarl==1.20.1
